Object Security - first working version + sample databases
git-svn-id: svn://localhost@267 f1514e4e-b68a-44c6-9dea-45284c435795

+++ trunk/core/src/com/sharifpro/eurb/management/mapping/dao/PersistableObjectDao.java
@@ -96,6 +96,6 @@ public interface PersistableObjectDao
 	 * @return list of objects
 	 * @throws PersistableObjectDaoException 
 	 */
-	public List<PersistableObject> findByPrimaryKeys(List<Integer> targetObjects) throws PersistableObjectDaoException;
+	public List<PersistableObject> findByPrimaryKeys(List<Long> targetObjects) throws PersistableObjectDaoException;
 
 }
